+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ### ? - ?
 
+##### Additions :tada:
+
+- Replaced the "Auto Fill" button on `CesiumFeaturesMetadataComponent` with "View Properties". This allows metadata properties and feature ID sets to be selectively added to the component, rather than adding them all in bulk.
+
 ##### Fixes :wrench:
 
 - Fixed a bug that prevented `UCesiumPrimitiveFeaturesBlueprintLibrary::GetPrimitiveFeatures` from retrieving the features of instanced meshes.